All Output calls now state who made the call, more polymorphism work

+++ src/main/java/config/InstagramSettings.java
@@ -1,7 +1,6 @@
 package config;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import services.HasUserID;
 
 import java.util.List;
 
@@ -34,7 +33,7 @@ public class InstagramSettings implements PlatformSettings {
     private boolean nsfw_allowed;
     private boolean use_reddit_caption;
     private List<String> caption_blacklist;
-    private long hours_before_duplicate_removed;
+    private int hours_before_duplicate_removed;
     private boolean audio_enabled;
     private String caption;
     private String hashtags;
@@ -62,8 +61,8 @@ public class InstagramSettings implements PlatformSettings {
     public int getAttempts_before_timeout() { return attempts_before_timeout; }
     public void setAttempts_before_timeout(int attempts_before_timeout) { this.attempts_before_timeout = attempts_before_timeout; }
 
-    public long getHours_before_duplicate_removed() { return hours_before_duplicate_removed; }
-    public void setHours_before_duplicate_removed(long hours_before_duplicate_removed) {this.hours_before_duplicate_removed = hours_before_duplicate_removed;}
+    public int getHours_before_duplicate_removed() { return hours_before_duplicate_removed; }
+    public void setHours_before_duplicate_removed(int hours_before_duplicate_removed) {this.hours_before_duplicate_removed = hours_before_duplicate_removed;}
 
     public List<String> getSubreddits() { return subreddits; }
     public void setSubreddits(List<String> subreddits) { this.subreddits = subreddits; }
