+++ src/main/java/config/YoutubeSettings.java
@@ -36,7 +36,7 @@ public class YoutubeSettings implements PlatformSettings,Refreshable {
     private boolean audio_enabled;
     private boolean nsfw_allowed;
     private boolean use_reddit_caption;
-    private long hours_before_duplicate_removed;
+    private int hours_before_duplicate_removed;
     private List<String> caption_blacklist;
     private String caption;
     private String hashtags;
@@ -59,8 +59,8 @@ public class YoutubeSettings implements PlatformSettings,Refreshable {
     public int getTime_between_posts() { return time_between_posts; }
     public void setTime_between_posts(int time_between_posts) { this.time_between_posts = time_between_posts; }
 
-    public long getHours_before_duplicate_removed() { return hours_before_duplicate_removed; }
-    public void setHours_before_duplicate_removed(long hours_before_duplicate_removed) {this.hours_before_duplicate_removed = hours_before_duplicate_removed;}
+    public int getHours_before_duplicate_removed() { return hours_before_duplicate_removed; }
+    public void setHours_before_duplicate_removed(int hours_before_duplicate_removed) {this.hours_before_duplicate_removed = hours_before_duplicate_removed;}
 
     public int getAttempts_before_timeout() { return attempts_before_timeout; }
     public void setAttempts_before_timeout(int attempts_before_timeout) { this.attempts_before_timeout = attempts_before_timeout; }
